@@ -28,17 +28,17 @@
 	"author": "Marko Zunic",
 	"license": "MIT",
 	"devDependencies": {
+		"@vitejs/plugin-vue": "^1.9.3",
+		"@vue/compiler-sfc": "^3.2.20",
 		"vite": "^2.6.10",
-		"vite-plugin-vue2": "^1.9.0",
-		"vue": "^2.6.14",
-		"vue-template-compiler": "^2.6.10"
+		"vue": "^3.2.20"
 	},
 	"peerDependencies": {
 		"moment": "^2.26.0"
 	},
 	"dependencies": {
-		"@jamescoyle/vue-icon": "^0.1.2",
 		"@mdi/js": "^6.3.95",
-		"lodash.throttle": "^4.1.1"
+		"lodash.throttle": "^4.1.1",
+		"vue3-icon": "^2.0.0"
 	}
 }
