@@ -1,8 +1,13 @@
-const path = require("path");
-const { createVuePlugin } = require("vite-plugin-vue2");
+import path from "path";
+import vue from "@vitejs/plugin-vue";
+import { defineConfig } from "vite";
 
-module.exports = {
-	plugins: [createVuePlugin()],
+export default defineConfig({
+	plugins: [
+		vue({
+			refTransform: true,
+		}),
+	],
 	build: {
 		lib: {
 			entry: path.resolve(__dirname, "src/index.js"),
@@ -18,4 +23,4 @@ module.exports = {
 			},
 		},
 	},
-};
+});
